Solved some bugs, made new ones.

+++ DRBN_STEAMVR2/Assets/SaveSnapShot.cs
@@ -461,17 +461,18 @@ public class SaveSnapShot : MonoBehaviour
 
         //File.WriteAllText("D:/trajectory_data/gamesave_list_test_"+ step.ToString() + ".jsonbrn", json);//not multiplatform, try something else 
         //File.WriteAllText("/mnt/d/trajectory_data/gamesave_list_test_" + step.ToString() + ".jsonbrn", json);//not multiplatform, try something else 
-        File.WriteAllText(Application.persistentDataPath + "/trajectory_data/gamesave_list_test_" + step.ToString() + ".jsonbrn", json);
-        Debug.Log("writing " + Application.persistentDataPath + "/trajectory_data/gamesave_list_test_" + step.ToString() + ".jsonbrn");
+        //File.WriteAllText(Application.persistentDataPath + "/trajectory_data/gamesave_list_test_" + step.ToString() + ".jsonbrn", json);
+        //Debug.Log("writing " + Application.persistentDataPath + "/trajectory_data/gamesave_list_test_" + step.ToString() + ".jsonbrn");
+
         //multiplatform but writing empty saves WTF
-        //if (Directory.Exists(Application.persistentDataPath + "/trajectory_data"))
-        //{
-        //    File.WriteAllText(Application.persistentDataPath + "/trajectory_data/gamesave_list_test_" + step.ToString() + ".jsonbrn", json);
-        //}
-        //else
-        //{
-        //    var folder = Directory.CreateDirectory(Application.persistentDataPath + "/trajectory_data");
-        //}
+        if (Directory.Exists(Application.persistentDataPath + "/trajectory_data"))
+        {
+            File.WriteAllText(Application.persistentDataPath + "/trajectory_data/gamesave_list_test_" + step.ToString() + ".jsonbrn", json);
+        }
+        else
+        {
+            var folder = Directory.CreateDirectory(Application.persistentDataPath + "/trajectory_data");
+        }
 
 
         //File.WriteAllText("/trajectory_data/gamesave_list_test_" + step.ToString() + ".jsonbrn", json);
